🚚 Rename: page -> pages로 폴더명 변경
